Moved DTO objects to DTO folder and verified new features still work after merging current version of main. Still need to add Account model to Invoice model
