+++ WellCoffeeInventoryAPI/src/main/java/com/team_lightyear/WellCoffeeInventoryAPI/dto/OrderedItemDTO.java
@@ -1,4 +1,4 @@
-package com.team_lightyear.WellCoffeeInventoryAPI.models;
+package com.team_lightyear.WellCoffeeInventoryAPI.dto;
 
 import java.util.Objects;
